% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_data.R
\name{all_markets}
\alias{all_markets}
\title{Get bids and asks for all 'PredictIt' markets}
\usage{
all_markets()
}
\value{
A \code{\link[=tibble]{tibble}} containing bid and ask data for all
  'PredictIt' markets.
}
\description{
Wrapper function to get all available 'PredictIt' markets and
  contract prices.
}
\examples{
markets <- all_markets()
markets

}
