% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_data.R
\name{format_market_data}
\alias{format_market_data}
\title{Format bid and ask market data with HTML}
\usage{
format_market_data(data)
}
\arguments{
\item{data}{Predictit market data, of class data.frame or tbl, as returned by pRedictit::get_predictit_market_data().}
}
\description{
Wrapper function to apply HTML formatting to Predictit market data. Can be displayed in a shiny app, or standalone in an htmlwidget (e.g. DT::datatable).
}
\examples{
data <- get_predictit_markets()
format_market_data(data)

}
