% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_data.R
\name{get_predictit_market}
\alias{get_predictit_market}
\title{Get bids and asks for a specific Predictit market}
\usage{
get_predictit_market(id)
}
\arguments{
\item{id}{Numerical code pertaining to the market. You can find a market's numerical code by consulting its URL or by first calling the all markets API.}
}
\description{
Wrapper function to get data for a specific market.
}
\examples{
markets <- get_predictit_markets()
id <- markets$id[1]
get_predictit_market(id)

}
