% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_data.R
\name{get_predictit_markets}
\alias{get_predictit_markets}
\title{Get bids and asks for all Predictit markets}
\usage{
get_predictit_markets()
}
\description{
Wrapper function to get all available Predictit markets and contract prices.
}
\examples{
markets <- get_predictit_markets()
markets

}
