% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_data.R
\name{get_predictit_markets_table}
\alias{get_predictit_markets_table}
\title{Get jQuery datatable containing bids and asks for all Predictit markets}
\usage{
get_predictit_markets_table()
}
\description{
Wrapper function to return a DT::datatable containing Predictit market data. Can be displayed in a shiny app, RMarkdown document, or exported via htmltools::saveWidget.
}
\examples{
get_predictit_markets_table()

}
