% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Run rpredictit examples}
\usage{
runExample(example)
}
\arguments{
\item{example}{The app to launch}
}
\description{
Launch a \code{rpredictit} example Shiny app that shows how to
easily use \code{rpredictit} in an app.\cr\cr
Run without any arguments to see a list of available example apps.
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
  # List all available example apps
  runExample()

  runExample("demo")
}
}
