% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_data.R
\name{tweet_markets}
\alias{tweet_markets}
\title{Get bids and asks for all PredictIt "Tweet" markets}
\usage{
tweet_markets()
}
\description{
Wrapper function to get all available PredictIt "Tweet count" markets and contract prices.
}
\examples{
tweet_markets()

}
