% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_data.R
\name{format_market_data}
\alias{format_market_data}
\title{Format bid and ask market data with HTML}
\usage{
format_market_data(data)
}
\arguments{
\item{data}{PredictIt market data, of class data.frame or tibble, as returned by predictit::all_markets() or predictit::tweet_markets()}
}
\description{
Wrapper function to apply HTML formatting to PredictIt market data. Can be displayed in a shiny app, or standalone in an htmlwidget (e.g. DT::datatable).
}
\examples{
data <- all_markets()
format_market_data(data)

}
