% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_data.R
\name{markets_table}
\alias{markets_table}
\title{Get JavaScript datatable containing bids and asks for all PredictIt markets}
\usage{
markets_table(data)
}
\arguments{
\item{data}{PredictIt market data, of class data.frame or tibble, as returned by rpredictit::all_markets() or rpredictit::tweet_markets()}
}
\description{
Wrapper function to return a DT::datatable containing PredictIt market data. Can be displayed in a shiny app, RMarkdown document, or exported via htmlwidgets::saveWidget.
}
\examples{
data <- all_markets()
markets_table(data)

}
